+++ src/main/scala/me/fornever/haskeletor/action/ShowTypeStickyAction.scala
@@ -11,9 +11,10 @@ package me.fornever.haskeletor.action
 import com.intellij.openapi.actionSystem.{ActionUpdateThread, AnAction, AnActionEvent}
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.PsiTreeUtil
+import me.fornever.haskeletor.core.util.StringUtil
 import me.fornever.haskeletor.external.component.HaskellComponentsManager
 import me.fornever.haskeletor.psi.{HaskellPsiUtil, HaskellQualifiedNameElement}
-import me.fornever.haskeletor.util.{HaskellEditorUtil, StringUtil}
+import me.fornever.haskeletor.util.HaskellEditorUtil
 
 import scala.annotation.tailrec