+++ src/main/scala/me/fornever/haskeletor/inspection/HLintInspectionTool.scala
@@ -15,7 +15,7 @@ import com.intellij.openapi.util.TextRange
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiFile
 import com.intellij.util.WaitFor
-import me.fornever.haskeletor.HaskellNotificationGroup
+import me.fornever.haskeletor.core.notifications.HaskellNotificationGroup
 import me.fornever.haskeletor.external.component.HLintRefactoringsParser._
 import me.fornever.haskeletor.external.component.{HLintComponent, HLintInfo, HLintRefactoringsParser}
 import me.fornever.haskeletor.util._