+++ src/main/scala/me/fornever/haskeletor/module/HaskellModuleBuilder.scala
@@ -19,12 +19,17 @@ import com.intellij.openapi.project.ex.ProjectManagerEx
 import com.intellij.openapi.projectRoots.SdkTypeId
 import com.intellij.openapi.roots._
 import com.intellij.openapi.roots.libraries.{Library, LibraryTablesRegistrar, LibraryUtil}
-import com.intellij.openapi.roots.ui.configuration.ModulesProvider
+import com.intellij.openapi.roots.ui.configuration.projectRoot.ProjectSdksModel
+import com.intellij.openapi.roots.ui.configuration.{ModulesProvider, SdkComboBox, SdkComboBoxModel}
 import com.intellij.openapi.ui.Messages
 import com.intellij.openapi.util.io.FileUtil
 import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.platform.templates.TemplateModuleBuilder
+import com.intellij.util.ui.JBUI
+import me.fornever.haskeletor.GlobalInfo
 import me.fornever.haskeletor.cabal.PackageInfo
+import me.fornever.haskeletor.core.HaskeletorBundle
+import me.fornever.haskeletor.core.notifications.HaskellNotificationGroup
 import me.fornever.haskeletor.external.component.{HaskellComponentsManager, LibraryPackageInfo}
 import me.fornever.haskeletor.external.execution.{CommandLine, StackCommandLine}
 import me.fornever.haskeletor.icons.HaskellIcons
@@ -32,11 +37,11 @@ import me.fornever.haskeletor.sdk.HaskellSdkType
 import me.fornever.haskeletor.settings.HaskellSettingsState
 import me.fornever.haskeletor.stackyaml.StackYamlComponent
 import me.fornever.haskeletor.util.{FutureUtil, HaskellFileUtil, HaskellProjectUtil, ScalaUtil}
-import me.fornever.haskeletor.{GlobalInfo, HaskellNotificationGroup}
 
+import java.awt.{GridBagConstraints, GridBagLayout}
 import java.io.File
 import java.nio.file.Path
-import javax.swing.Icon
+import javax.swing._
 import scala.annotation.tailrec
 import scala.collection.mutable.ListBuffer
 import scala.concurrent.duration._
@@ -164,20 +169,66 @@ class HaskellModuleBuilder extends TemplateModuleBuilder(null, HaskellModuleType
   }
 }
 
-class HaskellModuleWizardStep(wizardContext: WizardContext, haskellModuleBuilder: HaskellModuleBuilder) extends ProjectJdkForModuleStep(wizardContext, HaskellSdkType.getInstance) {
+class HaskellModuleWizardStep(wizardContext: WizardContext, haskellModuleBuilder: HaskellModuleBuilder) extends ModuleWizardStep {
+
+  private val sdkModel = SdkComboBoxModel.createSdkComboBoxModel(
+    wizardContext.getProject,
+    new ProjectSdksModel(),
+    sdkTypeId => sdkTypeId == HaskellSdkType.getInstance,
+    null,
+    null
+  )
+  private val sdkChooser = new SdkComboBox(sdkModel)
 
   override def updateDataModel(): Unit = {
-    super.updateDataModel()
-    haskellModuleBuilder.setModuleJdk(getJdk)
+    haskellModuleBuilder.setModuleJdk(sdkChooser.getSelectedSdk)
   }
 
   override def validate(): Boolean = {
-    if (getJdk == null) {
+    if (sdkChooser.getSelectedSdk == null) {
       Messages.showErrorDialog("You can't create a Haskell project without Stack configured as SDK", "No Haskell Tool Stack specified")
       false
     } else {
       true
     }
+  }
+
+  override def getComponent: JComponent = {
+    val panel = new JPanel(new GridBagLayout)
+    panel.setBorder(BorderFactory.createEtchedBorder())
+    panel.add(
+      new JLabel(HaskeletorBundle.message("module-wizard.sdk-chooser.label")),
+      new GridBagConstraints(
+        0,
+        0,
+        1,
+        1,
+        0.0,
+        0.0,
+        GridBagConstraints.WEST,
+        GridBagConstraints.NONE,
+        JBUI.insetsRight(5),
+        0,
+        0)
+    )
+
+    panel.add(
+      sdkChooser,
+      new GridBagConstraints(
+        0,
+        1,
+        1,
+        1,
+        1.0,
+        0.0,
+        GridBagConstraints.WEST,
+        GridBagConstraints.HORIZONTAL,
+        JBUI.insetsTop(5),
+        0,
+        0)
+    )
+
+    panel
   }
 }