+++ src/main/scala/me/fornever/haskeletor/psi/impl/HaskellPsiImplUtil.scala
@@ -13,9 +13,9 @@ import com.intellij.psi.impl.source.resolve.reference.ReferenceProvidersRegistry
 import com.intellij.psi.{PsiElement, PsiReference}
 import com.intellij.util.ArrayUtil
 import me.fornever.haskeletor.HaskellFileType
+import me.fornever.haskeletor.core.util.StringUtil
 import me.fornever.haskeletor.icons.HaskellIcons
 import me.fornever.haskeletor.psi._
-import me.fornever.haskeletor.util.StringUtil
 import org.jetbrains.annotations.Nullable
 
 import javax.swing._