+++ src/main/scala/me/fornever/haskeletor/sdk/HaskellSdkType.scala
@@ -15,13 +15,14 @@ import com.intellij.openapi.projectRoots.impl.SdkConfigurationUtil
 import com.intellij.openapi.roots.OrderRootType
 import com.intellij.openapi.util.SystemInfo
 import com.intellij.openapi.vfs.VirtualFile
-import me.fornever.haskeletor.HaskellNotificationGroup
+import me.fornever.haskeletor.core.notifications.HaskellNotificationGroup
 import me.fornever.haskeletor.external.execution.CommandLine
 import me.fornever.haskeletor.icons.HaskellIcons
 import me.fornever.haskeletor.util.{HaskellFileUtil, HaskellProjectUtil}
 import org.jdom.Element
 
 import java.io.File
+import java.nio.file.Path
 import javax.swing.Icon
 
 class HaskellSdkType extends SdkType("Haskell Tool Stack SDK") {
@@ -32,7 +33,7 @@ class HaskellSdkType extends SdkType("Haskell Tool Stack SDK") {
     else if (SystemInfo.isMac)
       "/usr/local/bin/stack"
     else if (SystemInfo.isWindows)
-      s"${sys.env.get("APPDATA")}\\local\\bin\\stack.exe"
+      sys.env.get("LOCALAPPDATA").map(Path.of(_).resolve("bin/stack.exe").toString).orNull
     else null
   }