+++ src/main/scala/me/fornever/haskeletor/util/ApplicationUtil.scala
@@ -13,7 +13,7 @@ import com.intellij.openapi.progress.ProgressIndicatorProvider
 import com.intellij.openapi.progress.util.ProgressIndicatorBase
 import com.intellij.openapi.project.Project
 import com.intellij.util.concurrency.AppExecutorUtil
-import me.fornever.haskeletor.HaskellNotificationGroup
+import me.fornever.haskeletor.core.notifications.HaskellNotificationGroup
 import me.fornever.haskeletor.external.component.{IndexNotReady, NoInfo, ReadActionTimeout}
 
 import java.util.concurrent.TimeUnit