+++ src/main/scala/me/fornever/haskeletor/util/FutureUtil.scala
@@ -9,7 +9,7 @@
 package me.fornever.haskeletor.util
 
 import com.intellij.openapi.project.Project
-import me.fornever.haskeletor.HaskellNotificationGroup
+import me.fornever.haskeletor.core.notifications.HaskellNotificationGroup
 
 import java.util.concurrent.{Future, TimeUnit, TimeoutException}